[stm32l475] upgrade lvgl porting
