+++ bsp/stm32/stm32l475-atk-pandora/board/ports/lvgl/lv_demo.c
@@ -10,10 +10,6 @@
 
 #include <rtthread.h>
 #include <lvgl.h>
-#include <drv_lcd.h>
-#define DBG_TAG    "LVGL"
-#define DBG_LVL    DBG_INFO
-#include <rtdbg.h>
 
 #ifndef LV_THREAD_STACK_SIZE
 #define LV_THREAD_STACK_SIZE 2048
@@ -26,51 +22,8 @@
 static void lv_example_get_started_1(void);
 static void lv_example_get_started_3(void);
 
-#define MY_DISP_HOR_RES 240 /* 240*240 */
-
-/*A static or global variable to store the buffers*/
-static lv_disp_draw_buf_t disp_buf;
-
-/*Static or global buffer(s). The second buffer is optional*/
-static lv_color_t buf_1[MY_DISP_HOR_RES * 10];
-static lv_color_t buf_2[MY_DISP_HOR_RES * 10];
-
-static lv_disp_drv_t disp_drv;  /*Descriptor of a display driver*/
-
-/*Flush the content of the internal buffer the specific area on the display
- *You can use DMA or any hardware acceleration to do this operation in the background but
- *'lv_disp_flush_ready()' has to be called when finished.*/
-static void disp_flush(lv_disp_drv_t * disp_drv, const lv_area_t * area, lv_color_t * color_p)
-{
-    lcd_fill_array(area->x1, area->y1, area->x2, area->y2, color_p);
-
-    /*IMPORTANT!!!
-     *Inform the graphics library that you are ready with the flushing*/
-    lv_disp_flush_ready(disp_drv);
-}
-
 static void lvgl_thread(void *parameter)
 {
-    lv_init();
-
-    /*Initialize `disp_buf` with the buffer(s). With only one buffer use NULL instead buf_2 */
-    lv_disp_draw_buf_init(&disp_buf, buf_1, buf_2, MY_DISP_HOR_RES*10);
-
-    lv_disp_drv_init(&disp_drv); /*Basic initialization*/
-
-    /*Set the resolution of the display*/
-    disp_drv.hor_res = MY_DISP_HOR_RES;
-    disp_drv.ver_res = MY_DISP_HOR_RES;
-
-    /*Set a display buffer*/
-    disp_drv.draw_buf = &disp_buf;
-
-    /*Used to copy the buffer's content to the display*/
-    disp_drv.flush_cb = disp_flush;
-
-    /*Finally register the driver*/
-    lv_disp_drv_register(&disp_drv);
-
     lv_example_get_started_1();
     lv_example_get_started_3();
 
@@ -90,7 +43,7 @@ static int lvgl_demo_init(void)
 
     return 0;
 }
-INIT_COMPONENT_EXPORT(lvgl_demo_init);
+INIT_APP_EXPORT(lvgl_demo_init);
 
 /* ------------------- demo1 ----------------------- */
