+++ bsp/stm32/stm32l475-atk-pandora/board/ports/lvgl/lv_port_indev.c
@@ -1,0 +1,14 @@
+/*
+ * Copyright (c) 2006-2021, RT-Thread Development Team
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * Change Logs:
+ * Date           Author       Notes
+ * 2021-10-18     Meco Man     The first version
+ */
+
+void lv_port_indev_init(void)
+{
+
+}
